requirements.txt
asgiref==3.3.1
attrs==20.3.0
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
coverage==5.3
cryptography==3.3.1
decorator==4.4.2
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1.4
django-bootstrap-datepicker-plus==3.0.5
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
dynaconf==3.1.2
flake8==3.8.4
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.8
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.3
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.1
pytest-cov==2.10.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
whitenoise==5.2.0
